+++ java/src/main/java/io/cucumber/gherkin/StringUtils.java
@@ -6,37 +6,89 @@ import static io.cucumber.gherkin.GherkinLanguageConstants.COMMENT_PREFIX_CHAR;
 class StringUtils {
 
     /**
-     * Matches regex pattern for whitespace.
+     * An extended definition of Whitespace minus new lines.
+     * <p>
+     * Character in Unicode general category {@code Zs} and directionality
+     * categories {@code WS}, {@code B}, and {@code S} are considered whitespace
+     * for this definition.
+     *
+     * @param c character to test
+     * @return true iff the {@code c} is whitespace and not new line..
      */
-    private static final char[] WHITESPACE_CHARS = new char[]{' ', '\t', '\n', '\u000B', '\f', '\r'};
+    private static boolean isWhiteSpaceExcludingNewLine(char c) {
+        return c != '\n' && isWhiteSpace(c);
+    }
 
     /**
-     * Matches regex pattern whitespace + NEL + NBSP.
+     * An extended definition of Whitespace.
+     * <p>
+     * Character in Unicode general category {@code Zs} and directionality
+     * categories {@code WS}, {@code B}, and {@code S} are considered whitespace
+     * for this definition.
+     *
+     * @param c character to test
+     * @return true iff the {@code c} is whitespace.
      */
-    private static final char[] WHITESPACE_CHARS_EXTENDED = new char[]{' ', '\t', '\n', '\u000B', '\f', '\r', '\u0085', '\u00A0'};
-    
-    /**
-     * Matches regex pattern whitespace + NEL + NBSP - new line.
-     */
-    private static final char[] WHITESPACE_CHARS_EXTENDED_KEEP_NEW_LINES = new char[]{' ', '\t', '\u000B', '\f', '\r', '\u0085', '\u00A0'};
+    static boolean isWhiteSpace(char c) {
+        // Fast path for the common case
+        return c == ' ' || c == '\t' || isCharacterTypeSpace(c) || isDirectionalitySpace(c);
+    }
+
+    private static boolean isCharacterTypeSpace(char c) {
+        return (((
+                (1 << Character.SPACE_SEPARATOR)
+                        // Not in the definition, but a subset of isDirectionalitySpace
+                        | (1 << Character.LINE_SEPARATOR)
+                        // Not in the definition, but a subset of isDirectionalitySpace
+                        | (1 << Character.PARAGRAPH_SEPARATOR)
+        ) >> Character.getType(c)) & 1) != 0;
+    }
+
+    private static boolean isDirectionalitySpace(char c) {
+        return (
+                (((1 << Character.DIRECTIONALITY_WHITESPACE)
+                        | (1 << Character.DIRECTIONALITY_PARAGRAPH_SEPARATOR)
+                        | (1 << Character.DIRECTIONALITY_SEGMENT_SEPARATOR)
+                ) >> Character.getDirectionality(c)) & 1) != 0;
+    }
+
     public static final IndentedText NO_INDENT_ENTRY = new IndentedText(0, "");
 
     static IndentedText trimAndIndentKeepNewLines(String input) {
-        return trimAndIndent(input, WHITESPACE_CHARS_EXTENDED_KEEP_NEW_LINES);
-    }
-
-    static IndentedText trimAndIndent(String input) {
-        return trimAndIndent(input, WHITESPACE_CHARS_EXTENDED);
-    }
-
-    private static IndentedText trimAndIndent(String input, char[] whitespaceChars) {
-        if (input.isEmpty()) {
+        int length = input.length();
+        if (length == 0) {
             return NO_INDENT_ENTRY;
         }
 
-        int start = findFirstIndexNotIn(input, input.length(), whitespaceChars);
-        int end = findLastIndexNotIn(input, start, whitespaceChars);
+        int start = 0;
+        while (start < length && isWhiteSpaceExcludingNewLine(input.charAt(start))) {
+            start++;
+        }
+        int end = length;
+        while (end > start && isWhiteSpaceExcludingNewLine(input.charAt(end - 1))) {
+            end--;
+        }
+        String trimmed = input.substring(start, end);
+        int indent = input.codePointCount(0, start);
+        // the object instance is not truly created because
+        // the code is inlined by the hotspot compiler
+        // (as "-XX:+EliminateAllocations" is enabled by default).
+        return new IndentedText(indent, trimmed);
+    }
 
+    static IndentedText trimAndIndent(String input) {
+        int length = input.length();
+        if (length == 0) {
+            return NO_INDENT_ENTRY;
+        }
+        int start = 0;
+        while (start < length && isWhiteSpace(input.charAt(start))) {
+            start++;
+        }
+        int end = length;
+        while (end > start && isWhiteSpace(input.charAt(end - 1))) {
+            end--;
+        }
         String trimmed = input.substring(start, end);
         int indent = input.codePointCount(0, start);
         // the object instance is not truly created because
@@ -50,7 +102,6 @@ class StringUtils {
      */
     static int findIndexOfTagComment(String input) {
         for (int i = 1, length = input.length(); i < length; i++) {
-            // TODO: Use isWhiteSpaceExtended for consistency instead?
             if (input.charAt(i) == COMMENT_PREFIX_CHAR && isWhiteSpace(input.charAt(i - 1))) {
                 return i - 1;
             }
@@ -58,42 +109,9 @@ class StringUtils {
         return -1;
     }
 
-    private static boolean isWhiteSpace(char c) {
-        return contains(WHITESPACE_CHARS, c);
-    }
-
-    private static int findFirstIndexNotIn(String input, int endIndex, char[] characters) {
-        int start = 0;
-        while (start < endIndex && contains(characters, input.charAt(start))) {
-            start++;
-        }
-        return start;
-    }
-
-    static boolean containsWhiteSpaceExtended(String input, int fromIndex, int toIndex) {
+    static boolean containsWhitespace(String input, int fromIndex, int toIndex) {
         for (int i = fromIndex; i < toIndex; i++) {
-            if (isWhiteSpaceExtended(input.charAt(i))) {
-                return true;
-            }
-        }
-        return false;
-    }
-    
-    static boolean isWhiteSpaceExtended(char input) {
-        return contains(WHITESPACE_CHARS_EXTENDED, input);
-    }
-    
-    private static int findLastIndexNotIn(String input, int beginIndex, char[] characters) {
-        int end = input.length();
-        while (end > beginIndex && contains(characters, input.charAt(end - 1))) {
-            end--;
-        }
-        return end;
-    }
-
-    private static boolean contains(char[] characters, char c) {
-        for (char candidate : characters) {
-            if (candidate == c) {
+            if (isWhiteSpace(input.charAt(i))) {
                 return true;
             }
         }